Refactor: 도로명 주소 검색창 로직 SearchBar 컴포넌트로 분리
